Move test_mons into test_services
To help group tests together and reduce the number of deployments needed.
